[70197] Move import/export file download class to a different directory
